Reworked object placing in the level editor, and may have fixed our missing links problems?
Before we maintained numbers for things like Player health, Robot speed, etc. and tried to parse them on each call to OnGUI. This meant that an exception would be thrown if the numerical text fields didn't have an appropriate format -- the empty string isn't an appropriate format, for example. This means that text editing was a pain in the editor. Now, these are all stored as strings and only parsed when we actually place the object. So basically nothing much changed except the interface sucks less WOO. I also added all the meta files -- go to Edit -> Project Settings -> Editor, and you should have version control set to "Meta Files".